v 1.1
The project pages on Coingecko have changed, so stage 3 has been removed from Workflow. At the moment, I have no plans to develop this robot, so the latest update makes the robot work, but if it were to be made professionally, it would have to be more adapted to the current Coingecko website. I don't plan on doing that though.